refactor: separate sqsClient and sqsConsumer
